+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 [Full changelog](https://github.com/mozilla/glean.js/compare/v0.17.0...main)
 
-* [#534](https://github.com/mozilla/glean.js/pull/534): Refactor Uploader into abstract class for external use
+* [#534](https://github.com/mozilla/glean.js/pull/534): Expose `Uploader` base class through `@mozilla/glean/<platform>/uploader` entry point.
 
 # v0.17.0 (2021-07-16)